requirements.txt
Flask==2.3.3               
flask-cors==3.0.10         
pymongo==3.11.0             
dnspython==1.16.0           
requests==2.31.0           
python-dotenv==1.1.1
solana==0.28.0
construct==2.10.67